tvm_ffi_sys

Module c_api

Structs

TVMFFIAny

TVM FFI Any value - a union type that can hold various data types

TVMFFIByteArray

Byte array data structure used by String and Bytes.

TVMFFIErrorCell

Error cell used in error object following header.

TVMFFIFieldInfo

Information support for optional object reflection

TVMFFIFunctionCell

Function cell

TVMFFIMethodInfo

Method information that can appear in reflection table

TVMFFIObject

TVMFFIShapeCell

Shape cell used in shape object following header.

TVMFFITypeAttrColumn

Column array that stores extra attributes about types

TVMFFITypeInfo

Runtime type information for object type checking

TVMFFITypeMetadata

Extra information of object type that can be used for reflection

Enums

TVMFFIBacktraceUpdateMode

TVMFFIObjectDeleterFlagBitMask

TVMFFITypeIndex

The index type of the FFI objects

Constants

COMBINED_REF_COUNT_BOTH_ONE

COMBINED_REF_COUNT_MASK_U32

COMBINED_REF_COUNT_STRONG_ONE

COMBINED_REF_COUNT_WEAK_ONE

Functions

TVMFFIAnyViewToOwnedAny⚠

TVMFFIBytesFromByteArray⚠

TVMFFIDataTypeFromString⚠

TVMFFIDataTypeToString⚠

TVMFFIErrorCreate⚠

TVMFFIErrorMoveFromRaised⚠

TVMFFIErrorSetRaised⚠

TVMFFIErrorSetRaisedFromCStr⚠

TVMFFIFunctionCall⚠

TVMFFIFunctionCreate⚠

TVMFFIFunctionGetGlobal⚠

TVMFFIFunctionSetGlobal⚠

TVMFFIGetTypeInfo⚠

TVMFFIStringFromByteArray⚠

TVMFFITensorFromDLPack⚠

TVMFFITensorFromDLPackVersioned⚠

TVMFFITensorToDLPack⚠

TVMFFITensorToDLPackVersioned⚠

TVMFFITestingDummyTarget⚠

TVMFFITraceback⚠

TVMFFITypeKeyToIndex⚠

Type Aliases

TVMFFIFieldGetter

Field getter function pointer type

TVMFFIFieldSetter

Field setter function pointer type

TVMFFIObjectCreator

Object creator function pointer type

TVMFFIObjectDeleter

TVMFFIObjectHandle

Handle to Object from C API’s pov

TVMFFISafeCallType

Safe call type for function ABI

Unions

TVMFFIAnyDataUnion

Second union in TVMFFIAny - 8 bytes